@@ -1,11 +1,11 @@
 {
   "name": "react-cool-starter",
-  "version": "2.2.0",
+  "version": "2.4.0",
   "description": "A starter boilerplate for an universal web application with the best development experience and best practices.",
   "main": "index.js",
   "engines": {
-    "node": ">=5.0 <8",
-    "npm": ">=3.0 <5"
+    "node": ">=6.0",
+    "npm": ">=3.0"
   },
   "repository": {
     "type": "git",
@@ -202,50 +202,52 @@
     "prop-types": "^15.5.10",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
-    "react-helmet": "^5.1.3",
+    "react-helmet": "^5.2.0",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-redux": "^5.0.5",
-    "react-router-config": "^1.0.0-beta.3",
-    "react-router-dom": "^4.1.1",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
     "react-router-redux": "5.0.0-alpha.6",
-    "redbox-react": "^1.4.2",
-    "redux": "^3.7.1",
+    "redbox-react": "^1.5.0",
+    "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "serialize-javascript": "^1.3.0",
+    "serialize-javascript": "^1.4.0",
     "serve-favicon": "^2.4.3",
     "webpack-isomorphic-tools": "^3.0.3"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.1",
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.1",
+    "autoprefixer": "^7.1.3",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "8.0.0",
+    "babel-jest": "^21.0.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-istanbul": "^4.1.4",
     "babel-preset-latest": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "better-npm-run": "0.0.15",
+    "babili-webpack-plugin": "^0.1.2",
+    "better-npm-run": "^0.1.0",
     "coveralls": "^2.13.1",
-    "css-loader": "^0.28.4",
+    "cross-spawn": "^5.1.0",
+    "css-loader": "^0.28.7",
     "enzyme": "^2.9.1",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-loader": "^1.8.0",
-    "eslint-plugin-import": "^2.6.1",
-    "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^2.1.2",
+    "eslint": "^4.6.1",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "^7.3.0",
+    "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.11.2",
-    "flow-bin": "^0.49.1",
-    "image-webpack-loader": "^3.3.1",
+    "flow-bin": "0.52.0",
+    "identity-obj-proxy": "^3.0.0",
+    "image-webpack-loader": "^3.4.1",
     "imports-loader": "^0.7.1",
-    "jest": "^20.0.4",
-    "nock": "^9.0.13",
+    "jest": "^21.0.1",
+    "nock": "^9.0.14",
     "node-sass": "^4.5.3",
-    "nodemon": "^1.11.0",
-    "postcss": "^6.0.4",
+    "nodemon": "^1.12.0",
+    "postcss": "^6.0.10",
     "postcss-loader": "^2.0.6",
     "react-addons-test-utils": "^15.6.0",
     "react-test-renderer": "^15.6.1",
@@ -253,13 +255,13 @@
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
-    "stylelint": "^7.12.0",
-    "stylelint-config-standard": "^16.0.0",
-    "stylelint-webpack-plugin": "^0.8.0",
+    "stylelint": "^8.1.1",
+    "stylelint-config-standard": "^17.0.0",
+    "stylelint-webpack-plugin": "^0.9.0",
     "url-loader": "^0.5.9",
-    "webpack": "^2.6.1",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-hot-middleware": "^2.18.0",
+    "webpack": "^3.5.5",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-hot-middleware": "^2.18.2",
     "webpack-node-externals": "^1.6.0"
   }
 }
