@@ -1,2 +1,3 @@
-// Return an object to emulate css modules (if you are using them)
-module.exports = {};
+import idObj from 'identity-obj-proxy';
+
+export default idObj;
