@@ -5,7 +5,7 @@ const WebpackIsomorphicTools = require('webpack-isomorphic-tools');
 // Setup global variables for server
 global.__CLIENT__ = false;
 global.__SERVER__ = true;
-global.__DISABLE_SSR__ = false;   // Disable server side render here
+global.__DISABLE_SSR__ = false; // Disable server side render here
 global.__DEV__ = process.env.NODE_ENV !== 'production';
 
 // This should be the same with webpack context
@@ -18,6 +18,6 @@ global.webpackIsomorphicTools = new WebpackIsomorphicTools(require('./tools/webp
       require('./src/server');
     } else {
       // $FlowFixMe: server.js will be generated during compiling
-      require('./build/server');  // eslint-disable-line import/no-unresolved
+      require('./build/server'); // eslint-disable-line import/no-unresolved
     }
   });
