@@ -19,12 +19,8 @@ export const fetchUser = (userId: string, axios: any, URL: string = API_URL): Th
     dispatch({ type: USER_REQUESTING, userId });
 
     return axios.get(`${URL}/${userId}`)
-      .then((res) => {
-        dispatch({ type: USER_SUCCESS, userId, data: res.data });
-      })
-      .catch((err) => {
-        dispatch({ type: USER_FAILURE, userId, err });
-      });
+      .then(res => dispatch({ type: USER_SUCCESS, userId, data: res.data }))
+      .catch(err => dispatch({ type: USER_FAILURE, userId, err: err.message }));
   };
 
 // Using for preventing dobule fetching data
