@@ -11,7 +11,7 @@ export default [
   {
     path: '/',
     exact: true,
-    component: HomePage,  // Add your route here
+    component: HomePage, // Add your route here
     loadData: (dispatch: Dispatch) => Promise.all([
       dispatch(fetchUsersIfNeeded()), // Register your server-side call action(s) here
     ]),
